refactor setting of multi maxchargecurrent
The idea here is to ensure that in systems with no BMS,
where a user current limit is specified, this is always
written to the Multi, not only when it changes, which can
cause issues if the Multi was off when the limit was changed.

But at the same time, if there is no user current limit, we don't
want to constantly max out the current, because that interferes
with any code that wants to do its own thing with the maximum
charge current, eg via Node Red.

So, when the charge current is unset, write a large value to
it once, then leave it alone. At all other times, just write
the required limit.

https://github.com/victronenergy/venus/issues/1496